validateLogin/requirements.txt
pyJwt
pymongo
bcrypt